Extract table header toggle into separate component
